First proposal to include ServiceLinks
Provides a simplified description of an update of a SERVICE LINK, in order to provide an updated schematic shape (not ROUTE POINTs) in case of rerouting. Only updated ServiceLink are expected to be provided here, the scheduled one being provided by a NeTEx feed. The SERVICE JOURNEY itself is knonw from hte context.
Note that, The shape (LineString) is mandatory in order to be consistent with the use case, which is to update the shape.
In this initial proposal, the Service Links update can be inserted either in a EstimatedVehicleJourneyStructure (to be repeated for each journey) either in a consequence of a PTSituation (can be shared accross multiple journeys). As usual, the EstimatedVehicleJourneyStructure  can be linked with the PTSituation.